refactor event handling to use int64 for UserIds and update token validation to return user ID

+++ model/event.go
@@ -12,7 +12,7 @@ type Event struct {
 	Description string    `bind:"required"`
 	Location    string    `bind:"required"`
 	Date        time.Time `bind:"required"`
-	UserIds     int
+	UserIds     int64
 }
 
 var events []Event
